# pointerify the rollingstate receiver